modif

+++ backend/controllers/book.js
@@ -16,11 +16,24 @@ exports.createThing = (req, res, next) => {
       .catch(error => res.status(400).json({ error }));
   };
 
-  exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-      .catch(error => res.status(400).json({ error }));
-  };
+  exports.deleteThing = (req, res, next ) => {
+    Thing.findOne({_id: req.params.id})
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message : 'Not authorized'});
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+               fs.unlink(`images/${filename}`, () => {
+                   Thing.deleteOne({_id: req.params.id})
+                       .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                       .catch(error => res.status(401).json({ error }));
+               });
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
+ }
 
   exports.getOneThing = (req, res, next) => {
     Thing.findOne({ _id: req.params.id })
@@ -33,3 +46,15 @@ exports.createThing = (req, res, next) => {
       .then(thing => res.status(200).json(thing))
       .catch(error => res.status(400).json({ error }));
   };
+
+
+
+
+
+
+
+
+
+
+
+
